refactor: Piece추상클래스로 변경
