+++ src/main/java/janggi/domain/piece/Elephant.java
@@ -6,7 +6,7 @@ import janggi.domain.piece.moveStrategy.MoveStrategy;
 import java.util.List;
 import java.util.Set;
 
-public class Elephant extends PieceAbstractInterface {
+public class Elephant extends Piece {
 
     private static final Set<List<Direction>> PATHS = Set.of(
             List.of(Direction.UP, Direction.UP_LEFT, Direction.UP_LEFT),
