+++ src/main/java/janggi/domain/piece/General.java
@@ -1,6 +1,6 @@
 package janggi.domain.piece;
 
-public class General extends PieceAbstractInterface {
+public class General extends Piece {
 
     public General() {
         super(null, null);
