+++ src/main/java/janggi/domain/piece/Guard.java
@@ -1,6 +1,6 @@
 package janggi.domain.piece;
 
-public class Guard extends PieceAbstractInterface {
+public class Guard extends Piece {
 
     public Guard() {
         super(null, null);
